Refactored the whole Repo
- Accepts Tabs and Stack mix navigation
