+++ app/(auth)/sign-in.tsx
@@ -58,7 +58,7 @@ export default function signIN() {
       if (signInAttempt.status == "complete") {
         await setActive({ session: signInAttempt.createdSessionId });
         console.log("Signed in Successfully");
-        router.replace("/");
+        router.replace("/home");
       } else {
         console.error(JSON.stringify(signInAttempt, null, 2));
       }
