+++ app/(tabs)/home/_layout.tsx
@@ -1,0 +1,10 @@
+
+import { Stack } from "expo-router";
+
+export default function HomeLayout() {
+    return (
+        <Stack>
+            <Stack.Screen name="index" options={{ headerShown: false }} />
+        </Stack>
+    );
+}
