+++ app/(tabs)/profile/_layout.tsx
@@ -1,0 +1,10 @@
+
+import { Stack } from "expo-router";
+
+export default function ProfileLayout() {
+    return (
+        <Stack>
+            <Stack.Screen name="index" options={{ headerShown: false }} />
+        </Stack>
+    );
+}
