+++ app/_layout.tsx
@@ -43,7 +43,7 @@ export default function RootLayout() {
           
           <Stack>
             <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-            <Stack.Screen name="(auth)" options={{ headerShown: false }} />
+            {/* <Stack.Screen name="(auth)" options={{ headerShown: false }} /> */}
             <Stack.Screen name="+not-found" />
             <Stack.Screen name="privacy-policy" options={{headerStyle: {
               backgroundColor: AppColors.OffWhite,
